Minor restructure, move to lsp_server module instead of server command
