+++ src/lib.rs
@@ -8,6 +8,7 @@ pub mod external;
 pub mod filesystem;
 pub mod imports;
 pub mod interfaces;
+pub mod lsp_server;
 pub mod modules;
 pub mod parsing;
 pub mod pattern;
@@ -116,8 +117,8 @@ impl From<sync::SyncError> for PyErr {
     }
 }
 
-impl From<server::ServerError> for PyErr {
-    fn from(err: server::ServerError) -> Self {
+impl From<lsp_server::error::ServerError> for PyErr {
+    fn from(err: lsp_server::error::ServerError) -> Self {
         PyOSError::new_err(err.to_string())
     }
 }
@@ -373,7 +374,7 @@ pub fn sync_project(
 fn run_server(
     project_root: PathBuf,
     project_config: config::ProjectConfig,
-) -> Result<(), server::ServerError> {
+) -> Result<(), lsp_server::error::ServerError> {
     server::run_server(project_root, project_config)
 }
