+++ src/lsp_server/mod.rs
@@ -1,0 +1,2 @@
+pub mod error;
+pub mod server;
